# openapi.yaml
...
openapi: 3.0.1
info:
  title: IDS AppStore
  description: IDS AppStore developed by the Fraunhofer FIT
  contact:
    name: IDS AppStore
    url: https://spring.io/projects/spring-boot/appstore
    email: '@email@'
  license:
    name: "Apache License, Version 2.0"
    url: https://www.apache.org/licenses/LICENSE-2.0.txt
  version: 1.0.0
servers:
  - url: https://localhost:8080
    description: Generated server url
...
paths:
...
  /api/notify:
    put:
      tags:
        - Messages
      summary: Notify all subscribers
      description: "Can be used to manually notify all subscribers about a resource\
        \ offer, representation, or artifact update."
      operationId: sendMessage
      parameters:
        - name: elementId
          in: query
          description: The element id.
          required: true
          schema:
            type: string
            format: uri
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                type: object
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: object
        "204":
          description: No Content
          content:
            '*/*':
              schema:
                type: object
...
  /api/webhook/registry:
    post:
      tags:
        - "Backend: Registry Webhooks"
      operationId: registryWebhook
      requestBody:
        content:
          application/json:
            schema:
              type: string
              description: The JSON Webhook event
        required: true
      responses:
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                type: object
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                type: object
...
  /api/apps/search:
    get:
      tags:
        - Data Apps
        - UI
      summary: Search all Apps with value in description
      operationId: searchApps
      parameters:
        - name: search
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 30
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                type: object
...